chore: rename files & remove extra code
