+++ Core/lambda/api_encoder.py
